[Mono] Move several small related files
