+++ modules/mono/glue/cs_files/Attributes/ToolAttribute.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace Godot
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class ToolAttribute : Attribute {}
+}
