--- modules/mono/glue/cs_files/IAwaiter.cs
@@ -1,18 +1,0 @@
-using System.Runtime.CompilerServices;
-
-namespace Godot
-{
-	public interface IAwaiter : INotifyCompletion
-	{
-		bool IsCompleted { get; }
-
-		void GetResult();
-	}
-
-	public interface IAwaiter<out TResult> : INotifyCompletion
-	{
-		bool IsCompleted { get; }
-
-		TResult GetResult();
-	}
-}
