--- modules/mono/glue/cs_files/ToolAttribute.cs
@@ -1,7 +1,0 @@
-using System;
-
-namespace Godot
-{
-	[AttributeUsage(AttributeTargets.Class)]
-	public class ToolAttribute : Attribute {}
-}
